Use roundtrip test instead of stringwriter
