+++ iidm/iidm-converter-api/src/test/resources/postProcessorReporterTest.json
@@ -1,0 +1,15 @@
+{
+  "version" : "1.0",
+  "reportTree" : {
+    "taskKey" : "testPostProcessor",
+    "reports" : [ {
+      "reportKey" : "testImportPostProcessor"
+    } ]
+  },
+  "dics" : {
+    "default" : {
+      "testPostProcessor" : "Test importer post processor",
+      "testImportPostProcessor" : "testing import post processor"
+    }
+  }
+}
