--- iidm/iidm-converter-api/src/test/resources/postProcessorReporterTest.txt
@@ -1,2 +1,0 @@
-+ Test importer post processor
-   testing import post processor
